Cleanup util (#22)
* Move utf8_decode_and_validate_uri_path_inplace to utf8_decoder

* Move urlencoded decode functions from util to urlencoded
